feat: Enable non-modular includes in framework modules
This commit enables the option to allow non-modular includes in framework modules in the Xcode project file. This is necessary for the project to build successfully.

Refactor the Xcode project file to include the `CLANG_ALLOW_NON_MODULAR_INCLUDES_IN_FRAMEWORK_MODULES` build setting with a value of `YES` for both the Debug and Release configurations.

ref: https://github.com/firebase/flutterfire/issues/13323